test(app): colocate the connector Chat round trip next to the composer
evals/specs now only holds specs that cross a product boundary (#4398,
#4402). The den-web link → desktop parser → composer chip round trip is a
unit contract, so it moves beside chat-link-parse.test.ts as a bun test and
the boundary ratchet is clean again.
